make a single patch to handle BottleRelics::onEquip
